Wrap body in RefCell
This allows us to pass an immutable reference to the request. Wrapping
objects that need mutability in RefCell has been previously done with
the request for form data and user data.

It is worth considering the cost of wrapping in RefCell. It may be
better to just pass a mutable reference to handlers.